+++ src/BlockLoader.php
@@ -36,13 +36,6 @@ class BlockLoader extends Singleton {
 	private $config;
 
 	/**
-	 * Flag that categories has been added.
-	 *
-	 * @var boolean
-	 */
-	private $categories_added = false;
-
-	/**
 	 * Headers to read from template files
 	 *
 	 * @var array
@@ -109,7 +102,7 @@ class BlockLoader extends Singleton {
 			$this->default_category = $this->categories[0]['slug'];
 		}
 
-		if ( class_exists( 'Micropackage\\ACFBlockCreator\\ACFBlockCreator' ) ) {
+		if ( class_exists( ACFBlockCreator::class ) ) {
 			$block_creator_config = array_filter( $this->config, function( $value, $key ) {
 				return in_array( $key, [
 					'blocks_dir',
@@ -338,31 +331,16 @@ class BlockLoader extends Singleton {
 	/**
 	 * Registers custom block category.
 	 *
-	 * `block_categories` filter is deprecated but has been left here for backward compatibility.
-	 *
 	 * @filter block_categories_all
-	 * @filter block_categories
 	 *
 	 * @since 1.0.0
-	 * @param array                             $categories Block categories.
-	 * @param \WP_Post|\WP_Block_Editor_Context $context    Block editor context or post object.
+	 * @param array                    $categories Block categories.
+	 * @param \WP_Block_Editor_Context $context    Block editor context.
 	 */
 	public function block_categories( $categories, $context ) {
-		$is_deprecated_hook = $context instanceof \WP_Post;
-
-		if (
-			$this->categories &&
-			( ! $is_deprecated_hook || false === $this->categories_added )
-		) {
+		if ( $this->categories ) {
 			$categories = array_merge( $categories, $this->categories );
 		}
-
-		/**
-		 * `block_categories_all` filter gets executed first, then deprecated `block_categories` gets executed if there is a
-		 * post object available. So we set the flag that the categories has been added already in the first case and reset it
-		 * in the other.
-		 */
-		$this->categories_added = ! $is_deprecated_hook;
 
 		return $categories;
 	}