+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Next]
+
+* [Removed] `block_categories` deprecated filter
+
 ## 1.0.5
 
 * [Fixed] Deprecated filter name being used